ui-core : refacto. use a single InputStatusIcon
